+++ R/0-Initialization.R
@@ -329,7 +329,7 @@ Tree.init= function(S){
 
   S$Sim$TimetoPrun_Tree= rep_len(FALSE,length(S$Sim$Cycle))
   S$Sim$TimetoPrun_Tree[S$Sim$Plot_Age%in%S$Parameters$Pruning_Age_Tree&
-                          S$Met_c$DOY%in%S$Parameters$date_pruning_Tree]= TRUE
+                          S$Met_c$DOY%in%S$Parameters$D_pruning_Tree]= TRUE
 
   S$Sim$WindSpeed_Tree= S$Met_c$WindSpeed
   S$Sim$TairCanopy_Tree= S$Met_c$Tair